Add extra verification before downloading dataset. Add requirement for deployment.

+++ main.py
@@ -1,6 +1,7 @@
 import pandas as pd
 import os
 from azure.storage.blob import BlobServiceClient
+import os
 
 connect_str = os.getenv('AZURE_STORAGE_CONNECTION_STRING')
 blob_service_client = BlobServiceClient.from_connection_string(connect_str)
@@ -19,13 +20,19 @@ file_names = [
 ]
 
 
+
 container_name = 'blob-csv-files'
 download_path = "data/"
 
 container_client = blob_service_client.get_container_client(container=container_name)
 
+# Verify if download_path directory exists, otherwise create it
+if not os.path.exists(download_path):
+    os.makedirs(download_path)
+
 for blob_name in file_names:
     download_file_path = f"{download_path}{blob_name}"
-    blob_client = container_client.get_blob_client(blob_name)
-    with open(download_file_path, "wb") as download_file:
-        download_file.write(blob_client.download_blob().readall())
+    if not os.path.exists(download_file_path):
+        blob_client = container_client.get_blob_client(blob_name)
+        with open(download_file_path, "wb") as download_file:
+            download_file.write(blob_client.download_blob().readall())
